auth-srp using SessionAuthenticator
